+++ activestorage/CHANGELOG.md
@@ -1,2 +1,5 @@
+*   Support passing `user_project` argument to GCS bucket to support GCS buckets with “Requester pays” enabled
+
+    *Tom Naessens*
 
 Please check [8-1-stable](https://github.com/rails/rails/blob/8-1-stable/activestorage/CHANGELOG.md) for previous changes.